@@ -1,20 +1,24 @@
 {
   "name": "georgio.xyz",
+  "private": true,
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "serve": "vite preview",
+    "preview": "vite preview",
     "deploy": "gcloud compute scp --recurse dist/* nathan:/var/www/georgio.xyz/html/"
   },
   "dependencies": {
-    "vue": "^3.2.26"
+    "vue": "^3.4.21"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^1.9.2",
-    "autoprefixer": "^10.4.1",
-    "postcss": "^8.4.5",
-    "tailwindcss": "^2.2.19",
-    "vite": "^2.7.10"
-  }
+    "@vitejs/plugin-vue": "^5.0.4",
+    "autoprefixer": "^10.4.19",
+    "cssnano": "^6.1.2",
+    "postcss": "^8.4.38",
+    "tailwindcss": "^3.4.1",
+    "vite": "^5.2.6"
+  },
+  "packageManager": "yarn@4.1.1"
 }
