@@ -1,6 +1,7 @@
-module.exports = {
-  purge: ["./index.html", "./src/**/*.{vue,js,ts,jsx,tsx}"],
-  darkMode: false, // or 'media' or 'class'
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: ["./index.html", "./src/**/*.{vue,js,ts,jsx,tsx}"],
+  darkMode: "media", // or 'media' or 'class'
   theme: {
     fontFamily: {
       sans: ["IBM Plex Sans"],
@@ -22,4 +23,5 @@ module.exports = {
     extend: {},
   },
   plugins: [],
-};
+}
+
